+++ urban-trails/src/components/Account.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Account = () => {
+    return (
+        <div>Account</div>
+    )
+}
+
+export default Account;
