+++ urban-trails/src/components/Signin.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Signin = () => {
+    return (
+        <div>Signin</div>
+    )
+}
+
+export default Signin;
