+++ urban-trails/src/components/Signup.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const Signup = () => {
+    return (
+        <div>Signup</div>
+    )
+}
+
+export default Signup
